fix: hydrate preferred language in cookie if cookie is not set
