+++ packages/bff/src/graphql/types/query.ts
@@ -1,5 +1,6 @@
 import { logger } from '@altinn/dialogporten-node-logger';
 import { list, objectType, stringArg } from 'nexus';
+import config from '../../config.js';
 import { SavedSearchRepository } from '../../db.ts';
 import {
   getNotificationAddressByOrgNumber,
@@ -7,7 +8,7 @@ import {
   getOrCreateProfile,
   getUserFromCore,
 } from '../functions/profile.ts';
-import { getLanguageFromAltinnContext } from './cookie.js';
+import { getLanguageFromAltinnContext, languageCodes, updateAltinnPersistentContextValue } from './cookie.js';
 import { getOrganizationsFromRedis } from './organization.ts';
 import { OrganizationResponse } from './profile.ts';
 
@@ -23,6 +24,23 @@ export const Query = objectType({
         const languageFromAltinnContext = getLanguageFromAltinnContext(
           ctx.request.raw.cookies?.altinnPersistentContext,
         );
+
+        // ensure the cookie uses the preferred language
+        if (!languageFromAltinnContext) {
+          const ul = languageCodes[language];
+          if (ul) {
+            const current = ctx.request.raw.cookies?.altinnPersistentContext;
+            const value = updateAltinnPersistentContextValue(current, ul);
+            ctx.request.context.reply.setCookie('altinnPersistentContext', value, {
+              path: '/',
+              domain: config.authContextCookieDomain,
+              expires: new Date(Date.now() + 1000 * 60 * 60 * 24 * 30),
+              httpOnly: true,
+              secure: true,
+            });
+          }
+        }
+
         return {
           language: languageFromAltinnContext || language,
           updatedAt,
